requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.12.3
python-dotenv==1.0.1
